+++ pyccel/parser/semantic.py
@@ -21,7 +21,7 @@ from sympy import ceiling
 from pyccel.utilities.strings import random_string
 from pyccel.ast.basic         import PyccelAstNode, TypedAstNode, ScopedAstNode
 
-from pyccel.ast.builtins import PythonPrint, PythonTupleFunction
+from pyccel.ast.builtins import PythonPrint, PythonTupleFunction, PythonSetFunction
 from pyccel.ast.builtins import PythonComplex
 from pyccel.ast.builtins import builtin_functions_dict, PythonImag, PythonReal
 from pyccel.ast.builtins import PythonList, PythonConjugate , PythonSet
@@ -64,7 +64,7 @@ from pyccel.ast.datatypes import CustomDataType, PyccelType, TupleType, VoidType
 from pyccel.ast.datatypes import PrimitiveIntegerType, HomogeneousListType, StringType, SymbolicType
 from pyccel.ast.datatypes import PythonNativeBool, PythonNativeInt, PythonNativeFloat
 from pyccel.ast.datatypes import DataTypeFactory, PrimitiveFloatingPointType
-from pyccel.ast.datatypes import InhomogeneousTupleType, HomogeneousTupleType
+from pyccel.ast.datatypes import InhomogeneousTupleType, HomogeneousTupleType, HomogeneousSetType
 from pyccel.ast.datatypes import PrimitiveComplexType, FixedSizeNumericType
 
 from pyccel.ast.functionalexpr import FunctionalSum, FunctionalMax, FunctionalMin, GeneratorComprehension, FunctionalFor
@@ -145,6 +145,12 @@ import pyccel.decorators as def_decorators
 
 errors = Errors()
 pyccel_stage = PyccelStage()
+
+type_container = {
+                   PythonTupleFunction : HomogeneousTupleType,
+                   PythonList : HomogeneousListType,
+                   PythonSetFunction : HomogeneousSetType
+                  }
 
 #==============================================================================
 
@@ -1920,7 +1926,7 @@ class SemanticParser(BasicParser):
                 raise errors.report(f"Unknown annotation base {base}\n"+PYCCEL_RESTRICTION_TODO,
                         severity='fatal', symbol=expr)
             rank = 1
-            if len(args) == 2 and args[1] is LiteralEllipsis():
+            if len(args) == 2 and args[1] is LiteralEllipsis() or len(args) == 1:
                 syntactic_annotation = args[0]
                 if not isinstance(syntactic_annotation, SyntacticTypeAnnotation):
                     pyccel_stage.set_stage('syntactic')
@@ -1928,10 +1934,8 @@ class SemanticParser(BasicParser):
                     pyccel_stage.set_stage('semantic')
                 internal_datatypes = self._visit(syntactic_annotation)
                 type_annotations = []
-                if dtype_cls is PythonTupleFunction:
-                    class_type = HomogeneousTupleType
-                elif dtype_cls is PythonList:
-                    class_type = HomogeneousListType
+                if dtype_cls in type_container :
+                    class_type = type_container[dtype_cls]
                 else:
                     raise errors.report(f"Unknown annotation base {base}\n"+PYCCEL_RESTRICTION_TODO,
                             severity='fatal', symbol=expr)